backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
transformers==4.28.1
torch==2.0.0
python-dotenv==1.0.0
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
html2text==2020.1.16
websockets==11.0.3
python-multipart==0.0.6
aiofiles==23.1.0 